remove old cmd functions and replace with new `run_args()`
