+++ crates/pacdef/src/backend/pip.rs
@@ -5,8 +5,8 @@ use anyhow::Context;
 use anyhow::Result;
 use serde_json::Value;
 
-use crate::backend::run_args;
-use crate::backend::run_args_for_stdout;
+use crate::cmd::run_args;
+use crate::cmd::run_args_for_stdout;
 use crate::prelude::*;
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
