--- crates/pacdef/src/backend/root.rs
@@ -1,3 +1,0 @@
-use anyhow::Result;
-use std::process::Command;
-
